Move Thread functions in task subfolder
